Put request for organism modifications
